Fix OAuth token without expiration time causing login to fail.
This will default the expiration time of an OAuth token passed to OSP to 365 days
to avoid a KeyError and therefor an Internal Server Error from showing up.

This can happen when an OAuth provider does not set a time limit on a token or
simply decides to not deliver that information for whatever reason. One such
provider is GitLab.

Fixes #226.